+++ src/pdm/core.py
@@ -17,12 +17,12 @@ import os
 import pkgutil
 import sys
 from datetime import datetime
+from functools import cached_property
 from pathlib import Path
 from tempfile import TemporaryDirectory
 from typing import TYPE_CHECKING, cast
 
 import tomlkit.exceptions
-from resolvelib import Resolver
 
 from pdm import termui
 from pdm.__version__ import __version__
@@ -30,7 +30,7 @@ from pdm.cli.options import ignore_python_option, no_cache_option, non_interacti
 from pdm.cli.utils import ArgumentParser, ErrorArgumentParser
 from pdm.compat import importlib_metadata
 from pdm.exceptions import PdmArgumentError, PdmUsageError
-from pdm.installers import InstallManager, Synchronizer
+from pdm.installers import InstallManager
 from pdm.models.repositories import BaseRepository, PyPIRepository
 from pdm.project import Project
 from pdm.project.config import Config
@@ -69,8 +69,6 @@ class Core:
 
     project_class = Project
     repository_class: type[BaseRepository] = PyPIRepository
-    resolver_class = Resolver
-    synchronizer_class = Synchronizer
     install_manager_class = InstallManager
 
     def __init__(self) -> None:
@@ -342,6 +340,35 @@ class Core:
                 self.ui.error(
                     f"Failed to load plugin {plugin.name}={plugin.value}: {e}",
                 )
+
+    @cached_property
+    def uv_cmd(self) -> list[str]:
+        from pdm.compat import importlib_metadata
+
+        self.ui.info("Using uv is experimental and might break due to uv updates.")
+        # First, try to find uv in Python modules
+        try:
+            importlib_metadata.distribution("uv")
+        except ModuleNotFoundError:
+            pass
+        else:
+            return [sys.executable, "-m", "uv"]
+        # Try to find it in the typical place:
+        if (uv_path := Path.home() / ".cargo/bin/uv").exists():
+            return [str(uv_path)]
+        # If not found, try to find it in PATH
+        import shutil
+
+        path = shutil.which("uv")
+        if path:
+            return [path]
+        # If not found, try to find in the bin dir:
+        if (uv_path := Path(sys.argv[0]).with_name("uv")).exists():
+            return [str(uv_path)]
+        raise PdmUsageError(
+            "use_uv is enabled but can't find uv, please install it first: "
+            "https://docs.astral.sh/uv/getting-started/installation/"
+        )
 
 
 def main(args: list[str] | None = None) -> None: