+++ src/pdm/installers/__init__.py
@@ -1,3 +1,5 @@
 # flake8: noqa
 from pdm.installers.manager import InstallManager
 from pdm.installers.synchronizers import Synchronizer
+from pdm.installers.base import BaseSynchronizer
+from pdm.installers.uv import UvSynchronizer