+++ src/pdm/project/core.py
@@ -26,7 +26,7 @@ from pdm.models.repositories import BaseRepository, LockedRepository
 from pdm.models.requirements import Requirement, parse_line, parse_requirement, strip_extras
 from pdm.models.specifiers import PySpecSet
 from pdm.project.config import Config, ensure_boolean
-from pdm.project.lockfile import Lockfile
+from pdm.project.lockfile import FLAG_INHERIT_METADATA, Lockfile
 from pdm.project.project_file import PyProject
 from pdm.utils import (
     cd,
@@ -47,8 +47,10 @@ if TYPE_CHECKING:
 
     from pdm.core import Core
     from pdm.environments import BaseEnvironment
+    from pdm.installers.base import BaseSynchronizer
     from pdm.models.caches import CandidateInfoCache, HashCache, WheelCache
     from pdm.models.candidates import Candidate
+    from pdm.resolver.base import Resolver
     from pdm.resolver.providers import BaseProvider
     from pdm.resolver.reporters import RichLockReporter
 
@@ -126,6 +128,10 @@ class Project:
     def lockfile(self) -> Lockfile:
         if self._lockfile is None:
             self._lockfile = Lockfile(self.root / self.LOCKFILE_FILENAME, ui=self.core.ui)
+            if self.config.get("use_uv"):
+                self._lockfile.default_strategies.discard(FLAG_INHERIT_METADATA)
+            if not self.config["strategy.inherit_metadata"]:
+                self._lockfile.default_strategies.discard(FLAG_INHERIT_METADATA)
         return self._lockfile
 
     def set_lockfile(self, path: str | Path) -> None:
@@ -290,6 +296,8 @@ class Project:
         from pdm.cli.commands.venv.backends import BACKENDS
 
         backend: str = self.config["venv.backend"]
+        if backend == "virtualenv" and self.config["use_uv"]:
+            backend = "uv"
         venv_backend = BACKENDS[backend](self, python)
         path = venv_backend.create(
             force=True,
@@ -462,7 +470,7 @@ class Project:
 
         import inspect
 
-        from pdm.resolver.providers import BaseProvider, get_provider
+        from pdm.resolver.providers import get_provider
 
         if env_spec is None:
             env_spec = (
@@ -477,16 +485,9 @@ class Project:
             try:
                 locked_repository = self.get_locked_repository(env_spec)
             except Exception:  # pragma: no cover
-                if for_install:
-                    raise
                 if strategy != "all":
                     self.core.ui.warn("Unable to reuse the lock file as it is not compatible with PDM")
 
-        if for_install:
-            assert locked_repository is not None
-            return BaseProvider(
-                locked_repository, direct_minimal_versions=direct_minimal_versions, locked_candidates={}
-            )
         provider_class = get_provider(strategy)
         params: dict[str, Any] = {}
         if strategy != "all":
@@ -841,3 +842,24 @@ class Project:
     @property
     def lock_targets(self) -> list[EnvSpec]:
         return [self.environment.allow_all_spec]
+
+    def get_resolver(self) -> type[Resolver]:
+        """Get the resolver class to use for the project."""
+        from pdm.resolver.resolvelib import RLResolver
+        from pdm.resolver.uv import UvResolver
+
+        if self.config.get("use_uv"):
+            return UvResolver
+        else:
+            return RLResolver
+
+    def get_synchronizer(self, quiet: bool = False) -> type[BaseSynchronizer]:
+        """Get the synchronizer class to use for the project."""
+        from pdm.installers import BaseSynchronizer, Synchronizer, UvSynchronizer
+        from pdm.installers.uv import QuietUvSynchronizer
+
+        if self.config.get("use_uv"):
+            return QuietUvSynchronizer if quiet else UvSynchronizer
+        if quiet:
+            return BaseSynchronizer
+        return getattr(self.core, "synchronizer_class", Synchronizer)