+++ tests/test_project.py
@@ -415,7 +415,7 @@ def test_preserve_log_file(project, pdm, tmp_path, mocker):
     all_logs = list(tmp_path.joinpath("logs").iterdir())
     assert len(all_logs) == 0
 
-    mocker.patch.object(project.core.synchronizer_class, "synchronize", side_effect=Exception)
+    mocker.patch("pdm.installers.Synchronizer.synchronize", side_effect=Exception)
     result = pdm(["add", "pytz"], obj=project)
     assert result.exit_code != 0
     install_log = next(tmp_path.joinpath("logs").glob("pdm-install-*.log"))